Remove startNode/endNode, use addNode instead
Remove startNode and endNode.

Use addNode instead, it takes a name and function as arguments, the function is called with a new Node and is expected to return a complete Node, with all the added components (See main.lua)

This function is called everytime a Node with the provided name is instantiated.

Fixes #11